move some functions into tools.py
