+++ autopark_quantiles.py
@@ -19,10 +19,13 @@ Options:
   --version     Show version.
 """
 import matplotlib.pyplot as plt
-import numpy as np
 import pandas as pd
 from docopt import docopt
-from tools import calculate_hyst
+from tools import (
+    calculate_hyst,
+    intervals,
+    get_no_small_intervals,
+)
 
 def load_schedules_actual_planned():
     result = []
@@ -128,39 +131,6 @@ def attach_wind_methods(data, threshold, window_size):
     data["quantile3"] = data.v_max.rolling("30min").quantile(0.70)
     data["quantile_park3"] = calculate_hyst(data["quantile3"], 40, 10)
     return data
-
-
-def intervals(data_column):
-    """ Determine the length of intervals between decisions
-    """
-    list_of_intervals = []
-    true = 0
-    false = 0
-    for unit in data_column:
-        if unit == False:
-            false += 1
-            if true != 0:
-                list_of_intervals.append(true)
-            true = 0
-        else:
-            true += 1
-            if false != 0:
-                list_of_intervals.append(false)
-            false = 0
-    return list_of_intervals
-
-
-def get_no_small_intervals(intv_list):
-    """Determine the length and number of intervals that are less than 10 and 20 minutes
-    """
-    small_intervals = []
-    medium_intervals = []
-    for i in np.array(intv_list):
-        if i <= 10:
-            small_intervals.append(i)
-        if 10 <= i <= 20:
-            medium_intervals.append(i)
-    return len(small_intervals)
 
 
 def plots(data, col, start_time, end_time, hyst_min, hyst_window, name, outfile_name):
